Optimisation: down to 2.904s for ms32k.
By changing the data type.

+++ fc-solve/source/game_type_limit.h
@@ -8,4 +8,4 @@
 // game_type_limit.h: define fcs_game_limit.
 #pragma once
 #include <stdint.h>
-typedef uint8_t fcs_game_limit;
+typedef uint16_t fcs_game_limit;
